refactor(domain): clean architecture
