+++ app/src/main/java/com/kristallik/jokeapp/domain/usecase/GetNetworkJokesUseCase.kt
@@ -1,0 +1,14 @@
+package com.kristallik.jokeapp.domain.usecase
+
+import com.kristallik.jokeapp.data.model.NetworkJoke
+import com.kristallik.jokeapp.data.repository.JokeRepository
+
+class GetNetworkJokesUseCase(private val jokeRepository: JokeRepository) {
+    suspend fun execute(): List<NetworkJoke> {
+        return try {
+            jokeRepository.getAllNetworkJokes()
+        } catch (e: Exception) {
+            throw e
+        }
+    }
+}
